move genesis logic to keeper
