+++ x/twap/keeper.go
@@ -40,3 +40,30 @@ func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
 func (k *Keeper) PruneEpochIdentifier(ctx sdk.Context) string {
 	return k.GetParams(ctx).PruneEpochIdentifier
 }
+
+// InitGenesis initializes the capability module's state from a provided genesis
+// state.
+func (k Keeper) InitGenesis(ctx sdk.Context, genState *types.GenesisState) {
+	if err := genState.Validate(); err != nil {
+		panic(err)
+	}
+
+	k.SetParams(ctx, genState.Params)
+
+	for _, twap := range genState.Twaps {
+		k.storeNewRecord(ctx, twap)
+	}
+}
+
+// ExportGenesis returns the capability module's exported genesis.
+func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
+	twapRecords, err := k.getAllHistoricalPoolIndexedTWAPs(ctx)
+	if err != nil {
+		panic(err)
+	}
+
+	return &types.GenesisState{
+		Params: k.GetParams(ctx),
+		Twaps:  twapRecords,
+	}
+}
